langchain_adapter/pyproject.toml
[project]
name = "langchain-adapter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ddgs>=9.5.5",
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.12.0",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.21",
    "langchain-openai>=0.3.32",
    "langchain[google-genai]>=0.3.27",
    "langgraph>=0.6.6",
    "pydantic-settings>=2.10.1",
]
